Move ActionDensity to new file.
