+++ supervillain/observable/__init__.py
@@ -2,5 +2,5 @@ from .observable import Observable
 from .derived import DerivedQuantity
 
 from .energy import InternalEnergyDensity, InternalEnergyDensitySquared
-from .energy import ActionDensity
+from .action import ActionDensity
 from .topology import WindingSquared, Winding_Winding
